Separate out the layer manager to forms and hotspot
