+++ assets/src/js/godam-player/managers/layers/README.md
@@ -1,0 +1,22 @@
+# Layers Manager Components
+
+This directory contains the refactored components of the LayersManager class, split into focused, maintainable modules.
+
+## Structure
+
+- **`LayerValidator.js`** - Handles layer dependency checking and validation logic
+- **`FormLayerManager.js`** - Manages form layer functionality including skip buttons, form observation, and time-based updates
+- **`HotspotLayerManager.js`** - Manages hotspot layer functionality including creation, positioning, tooltip management, and interactions
+- **`../layersManager.js`** - Main orchestrator that coordinates the above components
+
+## Benefits of this refactoring
+
+1. **Single Responsibility Principle**: Each module has a clear, focused purpose
+2. **Better Maintainability**: Smaller files are easier to understand and modify
+3. **Improved Testability**: Individual components can be tested in isolation
+4. **Enhanced Reusability**: Components can potentially be reused in other contexts
+5. **Reduced Complexity**: The original 713-line file is now split into manageable chunks
+
+## Public Interface
+
+The main `LayersManager` class maintains the same public interface to ensure backward compatibility with existing code that depends on it.
